20. Gun Odev Commit

+++ Entities/Concrete/Car.cs
@@ -16,7 +16,7 @@ namespace Entities.Concrete
         public int BrandId { get; set; }
         public int ColorId { get; set; }
         public string ModelYear { get; set; }
-        public decimal DealyPrice { get; set; }
+        public int DealyPrice { get; set; }
         public string Description { get; set; }
         public List<CarImage> CarImages { get; set; }
