+++ Entities/Dtos/CarDetailDto.cs
@@ -14,7 +14,7 @@ namespace Entities.Dtos
         public string Brand { get; set; }
         public string Color { get; set; }
         public string ModelYear { get; set; }
-        public decimal DealyPrice { get; set; }
+        public int DealyPrice { get; set; }
         public string Description { get; set; }
         public List<CarImage> CarImages { get; set; }
     }
